Unlock function in progress

+++ Profile/ProfilePics/OtherProfilePicsViewController.swift
@@ -12,6 +12,8 @@ import SwiftyJSON
 private let ProfilePicCellIdentifier = "OtherProPicCell"
 private let ProfilePicCellNibName = "OtherProfilePicCollectionViewCell"
 
+private let picUnlockCost: Double = 50
+
 class OtherProfilePicsViewController: ProfilePicsViewController,
                                         LMCollectionViewLayoutDelegate,
                                         UICollectionViewDataSource
@@ -28,17 +30,78 @@ class OtherProfilePicsViewController: ProfilePicsViewController,
     @IBOutlet weak var nameLabel: UILabel!
     @IBOutlet weak var ageLabel: UILabel!
 
+    let unlockAlert = UIAlertController(title: "相片解锁", message: "您需要花费 \(picUnlockCost) P币来解锁该等级", preferredStyle: UIAlertControllerStyle.Alert)
+    let loadingAlert = UIAlertController(title: "载入中", message: "请稍后", preferredStyle: UIAlertControllerStyle.Alert)
 
+    var unlockId = 0
+
+    var userId: Int!
     var tenUser: TenUser!
 
     var pVC: OtherProfileViewController!
 
+    var unlocksJSON: [AnyObject]!
+
+
     override func viewDidLoad() {
         super.viewDidLoad()
 
+        navigationController?.presentViewController(loadingAlert, animated: true, completion: nil)
+
+        let targetUrl = Url_Unlocker + "?owner=\(userId)&unlocker=\(SHARED_USER.UserIndex)"
+
+        ALAMO_MANAGER.request(.GET, targetUrl) .responseJSON {response in
+
+            if let values = response.result.value {
+
+                self.unlocksJSON = (values as? [AnyObject])!
+                self.loadingAlert.dismissViewControllerAnimated(true, completion: nil)
+            }
+        }
+
+        let ok = UIAlertAction(title: "解锁", style: UIAlertActionStyle.Destructive, handler: { (ac) -> Void in
+            print("解锁")
+
+            if SHARED_USER.PCoin < picUnlockCost {
+                let insufficientAlert = UIAlertController(title: "解锁失败", message: "P币数量不足，请充值后解锁", preferredStyle: .Alert)
+                let pay = UIAlertAction(title: "充值", style: .Destructive, handler: { (ac) -> Void in
+                    let pVC = PCoinViewController()
+                    self.navigationController?.navigationBar.hidden = false
+                    self.navigationController?.pushViewController(pVC, animated: true)
+                })
+                let cancel = UIAlertAction(title: "取消", style: .Cancel, handler: nil)
+                insufficientAlert.addAction(pay)
+                insufficientAlert.addAction(cancel)
+                self.presentViewController(insufficientAlert, animated: true, completion: nil)
+            }
+            else {
+                let params: [String : AnyObject] =
+                [
+                    "TenImageID": self.unlockId,
+                    "Owner": self.userId,
+                    "Unlocker": SHARED_USER.UserIndex,
+                    "Pcoin": picUnlockCost,
+                    "UnlockTime": NSDate().timeIntervalSince1970
+                ]
+
+                ALAMO_MANAGER.request(.POST, Url_Unlocker, parameters: params, encoding: .JSON, headers: nil) .responseJSON {
+                    response in
+                    if response.result.isSuccess {
+                        SHARED_USER.PCoin -= picUnlockCost
+                    }
+                }
+            }
+        })
+
+        let cancel = UIAlertAction(title: "取消", style: UIAlertActionStyle.Cancel, handler: nil)
+        unlockAlert.addAction(cancel)
+        unlockAlert.addAction(ok)
+
+        //
+
         lmCollectionView.reloadData()
 
-        
+
         // register nib
         let cellNib = UINib(nibName: ProfilePicCellNibName, bundle: nil)
         lmCollectionView.registerNib(cellNib, forCellWithReuseIdentifier: ProfilePicCellIdentifier)
@@ -79,7 +142,12 @@ class OtherProfilePicsViewController: ProfilePicsViewController,
     }
     
     func collectionView(collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return self.dims.count
+        if let jsons = self.imagesJSON {
+            return jsons.count
+        }
+        else {
+            return 0
+        }
     }
 
     func collectionView(collectionView: UICollectionView, cellForItemAtIndexPath indexPath: NSIndexPath) -> UICollectionViewCell {
@@ -91,9 +159,23 @@ class OtherProfilePicsViewController: ProfilePicsViewController,
         let imageName = imageJSON["FileName"].stringValue
         let cachedImage = SHARED_IMAGE_CACHE.imageWithIdentifier(imageName)
 
+        unlockId = imageJSON["ID"].intValue
+
         if let retrivedImage = cachedImage {
             cell.picIV.image = retrivedImage
-            if imageJSON["IsLocked"].boolValue == true {
+
+            var unlockedByUser = false
+
+            for item in self.unlocksJSON! {
+                let unlockJSON = JSON(item)
+                print("UnlockedJSON: \(unlockJSON["TenImageID"].intValue) unlockId: \(unlockId)")
+                if unlockJSON["TenImageID"].intValue == unlockId {
+                    unlockedByUser = true
+                }
+                
+            }
+
+            if imageJSON["IsLocked"].boolValue == true && !unlockedByUser {
                 setLockStatus(cell, status: true)
             }
             else {
@@ -111,7 +193,18 @@ class OtherProfilePicsViewController: ProfilePicsViewController,
                     cell.picIV.image = image
 
                     SHARED_IMAGE_CACHE.addImage(image, withIdentifier: imageName)
-                    if imageJSON["IsLocked"].boolValue == true {
+
+                    var unlockedByUser = false
+                    for item in self.unlocksJSON! {
+                        let unlockJSON = JSON(item)
+                        print("UnlockedJSON: \(unlockJSON["TenImageID"].intValue) unlockId: \(self.unlockId)")
+                        if unlockJSON["TenImageID"].intValue == self.unlockId {
+                            unlockedByUser = true
+                        }
+
+                    }
+
+                    if imageJSON["IsLocked"].boolValue == true && !unlockedByUser {
                         self.setLockStatus(cell, status: true)
                     }
                     else {
@@ -137,35 +230,56 @@ class OtherProfilePicsViewController: ProfilePicsViewController,
 
         self.isProcessing = true
 
-        let selectedCell = lmCollectionView.cellForItemAtIndexPath(indexPath) as? OtherProfilePicCollectionViewCell
-        let cellImage = selectedCell?.picIV.image
+        var unlockedByUser = false
 
-        self.lmCollectionView.performBatchUpdates({() in
-            if self.dims[indexPath.row] == BlockDim.Std {
-                self.dims.removeAtIndex(indexPath.row)
-                self.lmCollectionView?.deleteItemsAtIndexPaths([indexPath])
+        let obj = self.imagesJSON[indexPath.row]
+        let imageJSON = JSON(obj as! [String: AnyObject])
 
-                self.dims.insert(BlockDim.L, atIndex: indexPath.row)
-                self.lmCollectionView?.insertItemsAtIndexPaths([indexPath])
+        unlockId = imageJSON["ID"].intValue
+
+        for item in self.unlocksJSON! {
+            let unlockJSON = JSON(item)
+            if unlockJSON["TenImageID"].intValue == unlockId {
+                unlockedByUser = true
             }
-            else {
-                self.dims.removeAtIndex(indexPath.row)
-                self.lmCollectionView?.deleteItemsAtIndexPaths([indexPath])
 
-                self.dims.insert(BlockDim.Std, atIndex: indexPath.row)
-                self.lmCollectionView?.insertItemsAtIndexPaths([indexPath])
-            }
-            },
-            completion: {(done) in
-                let novaCell = self.lmCollectionView.cellForItemAtIndexPath(indexPath) as? OtherProfilePicCollectionViewCell
-                novaCell?.picIV.image = cellImage
-                self.isProcessing = false
-        })
+        }
+
+        if imageJSON["IsLocked"] == true && !unlockedByUser {
+            self.navigationController?.presentViewController(unlockAlert, animated: true, completion: nil)
+
+        }
+        else {
+            let selectedCell = lmCollectionView.cellForItemAtIndexPath(indexPath) as? OtherProfilePicCollectionViewCell
+            let cellImage = selectedCell?.picIV.image
+
+            self.lmCollectionView.performBatchUpdates({() in
+                if self.dims[indexPath.row] == BlockDim.Std {
+                    self.dims.removeAtIndex(indexPath.row)
+                    self.lmCollectionView?.deleteItemsAtIndexPaths([indexPath])
+
+                    self.dims.insert(BlockDim.L, atIndex: indexPath.row)
+                    self.lmCollectionView?.insertItemsAtIndexPaths([indexPath])
+                }
+                else {
+                    self.dims.removeAtIndex(indexPath.row)
+                    self.lmCollectionView?.deleteItemsAtIndexPaths([indexPath])
+
+                    self.dims.insert(BlockDim.Std, atIndex: indexPath.row)
+                    self.lmCollectionView?.insertItemsAtIndexPaths([indexPath])
+                }
+                },
+                completion: {(done) in
+                    let novaCell = self.lmCollectionView.cellForItemAtIndexPath(indexPath) as? OtherProfilePicCollectionViewCell
+                    novaCell?.picIV.image = cellImage
+                    self.isProcessing = false
+            })
+        }
     }
 
     private func setLockStatus(cell: OtherProfilePicCollectionViewCell, status: Bool) {
         if status {
-            cell.picIV.alpha = 0.3
+            cell.picIV.alpha = 0.1
         }
         else {
             cell.picIV.alpha = 1.0
